requirements.txt
torch
transformers
datasets
peft
accelerate
matplotlib
pyyaml

